requirements.txt
openai
google-generativeai
python-dotenv
flask